feat: drop bin_saves, move SaveHeader
