+++ src/build-nc-w64/helapordo_win.h
@@ -34,7 +34,6 @@
 #include "floor_tester.h"
 
 #include "../utils/rooms.h"
-#include "../utils/bin_saves.h"
 #include "../anvil__helapordo.h"
 #include "../core/game_lore.h"
